% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_bootstrap.R
\name{mm_bootstrap}
\alias{mm_bootstrap}
\alias{mm_resample}
\alias{mm_boot_estimates}
\title{Generate bootstrap estimates of overlap}
\usage{
mm_bootstrap(
  A,
  B,
  nb,
  smooth = TRUE,
  kmax = 3,
  adjust = NA,
  n_grid = 128,
  type = c("Dhat1", "Dhat4", "Dhat5"),
  cores = 1
)

mm_resample(x, nb, smooth = TRUE, kmax = 3, adjust = 1, n_grid = 512)

mm_boot_estimates(
  Amat,
  Bmat,
  kmax = 3,
  adjust = c(0.8, 1, 4),
  n_grid = 128,
  type = c("all", "Dhat1", "Dhat4", "Dhat5"),
  cores = 1
)
}
\arguments{
\item{A}{A numeric vector of time-of-day observations (in radians) for species A.}

\item{B}{A numeric vector of time-of-day observations (in radians) for species B.}

\item{nb}{
the number of bootstrap samples required
}

\item{smooth}{
if TRUE, smoothed bootstrap samples are produced.
}

\item{kmax}{An integer indicating the maximum number of modes allowed in the activity pattern. Default is \code{3}.}

\item{adjust}{A numeric value to adjust the bandwidth of the kernel density estimation. Default is \code{1}.}

\item{n_grid}{An integer specifying the number of grid points for density estimation. Default is \code{128}.}

\item{type}{the name of the estimator to use, or "all" to produce all three
estimates. See \code{\link[=mm_overlap_estimates]{mm_overlap_estimates()}} for recommendations on which to use.}

\item{cores}{
the number of cores to use for parallel processing. If NA, all but one of the available cores will used. Parallel processing may take longer than serial processing if the bootstrap runs quickly.
}

\item{x}{
a numeric vector of time-of-capture data in \emph{radians}, ie. on [0, \eqn{2\pi}] scale
}

\item{Amat, Bmat}{
matrices of resampled data for each species produced by \code{resample}; see Value below.
}
}
\value{
The function \code{\link[=mm_bootstrap]{mm_bootstrap()}} returns a vector of bootstrap estimates.
If estimation fails for a bootstrap sample, the corresponding value will be NA.

The function \code{\link[=mm_resample]{mm_resample()}} returns a numeric matrix with each column corresponding
to a bootstrap sample. Times are in \code{radians}. It may return a matrix of NAs if
\code{smooth = TRUE} and bandwidth estimation fails.

The Function \code{\link[=mm_boot_estimates]{mm_boot_estimates()}} with \code{type = "all"} returns a numeric matrix
with three columns, one for each estimator of overlap, otherwise a vector of
bootstrap estimates.
}
\description{
\code{bootstrap} takes two sets of times of observations and calculates bootstrap
estimates of the chosen estimator of overlap. Alternatively, bootstrap
estimates can be calculated in a 2-stage process: (1) create a matrix of
bootstrap samples for each data set, using \code{\link[=mm_resample]{mm_resample()}}; (2) pass these
matrices to \code{\link[=mm_boot_estimates]{mm_boot_estimates()}} to obtain the bootstrap estimates.

A vector of bootstrap estimates can then be used to produce confidence intervals
with \code{\link[=mm_boot_ci]{mm_boot_ci()}}.
}
\examples{

# Generate random data for two species
set.seed(42)
species_A <- runif(100, 1.2, 2 * pi)
species_B <- runif(100, 0.23, 2 * pi)

est <- mm_overlap_estimates(species_A, species_B, type="Dhat4")

boots <- mm_bootstrap(species_A, species_B, 1000, type="Dhat4", cores=1)
mean(boots)
hist(boots)
mm_boot_ci(est, boots)

# alternatively:
species_A_gen <- mm_resample(species_A, 1000)
species_B_gen <- mm_resample(species_B, 1000)
boots <- mm_boot_estimates(species_A_gen, species_B_gen, type="Dhat4", cores=1)
mean(boots)

}
\references{
Ridout & Linkie (2009) Estimating overlap of daily activity patterns
from camera trap data. Journal of Agricultural, Biological, and Environmental
Statistics 14:322-337
}
\seealso{
\code{\link[=mm_boot_ci]{mm_boot_ci()}}
}
\author{
Mike Meredith, Martin Ridout.
}
