% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{camtrapdp}
\alias{camtrapdp}
\title{Camera trap data package example}
\format{
A list holding study data and metadata.
}
\usage{
camtrapdp
}
\description{
Data and metadata from an example study exported from the Agouti camera trap
data management platform in camtrap-DP format. Metadata includes study name,
authors, location and other details. Data is held in element data, itself a
list holding dataframes deployments, media and observations.
See \href{https://camtrap-dp.tdwg.org/}{https://tdwg.github.io/camtrap-dp} for details.
}
\author{
Marcus Rowcliffe
}
\keyword{datasets}
