% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_check_name.R
\name{mm_check_name}
\alias{mm_check_name}
\title{Check species name and retrieve Taxonomic Serial Number (TSN) from ITIS}
\usage{
mm_check_name(species_name, search_type, ask = FALSE)
}
\arguments{
\item{species_name}{A character string specifying the species name to search
for. Only a \strong{single name} is allowed.}

\item{search_type}{A character string specifying the type of search. Options:
\itemize{
\item \code{"sientific_name"}: Search by scientific name.
\item \code{"common_name"}: Search by common name.
}}

\item{ask}{A logical value (\code{TRUE} or \code{FALSE}). If \code{TRUE}, allows interactive selection when multiple matches are found.}
}
\value{
A tibble containing taxonomic details:
\itemize{
\item \code{search}: The original species name queried.
\item \code{tsn}: The \strong{Taxonomic Serial Number} (TSN) from ITIS.
\item \code{common_name}: The common name of the species (if available).
\item \code{scientific_name}: The scientific name of the species.
\item \code{author}: The author who classified the species.
\item \code{itis_url}: A direct link to the species report on ITIS.
\item \code{taxon_status}: The taxonomic status of the species.
}
}
\description{
This function queries the \strong{Integrated Taxonomic Information System (ITIS)}
to find taxonomic details for a given species name.  It can search using
either a \strong{scientific name} or a \strong{common name} and return relevant
taxonomic information, including the TSN.
}
\details{
\itemize{
\item If the necessary packages (\code{httr2}, \code{xml2}) are not installed, the function prompts the user to install them.
\item If multiple results are found and \code{ask = TRUE}, the user is prompted to select the correct match.
\item If no exact match is found, all results are displayed for manual selection.
}
}
\examples{
\dontrun{
# Search for a species by scientific name
mm_check_name("Panthera leo", search_type = "sientific_name")

# Search by common name with interactive selection
mm_check_name("Lion", search_type = "common_name", ask = TRUE)
}


}
\seealso{
\url{https://www.itis.gov}
}
