% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_describe_df.R
\name{mm_describe_df}
\alias{mm_describe_df}
\title{Descriptive statistic on dataset}
\usage{
mm_describe_df(data, ..., fn = NULL)
}
\arguments{
\item{data}{A data frame containing the dataset to be summarized.}

\item{...}{(Optional) Column names to include in the summary. If no columns
are specified, all columns in the dataset will be included.}

\item{fn}{A named list of functions to apply to numeric variables. Each function
must accept \code{x} as a vector of numeric values and return a single value or a
named vector. Additional arguments for these functions can be specified as a list.
For example: \code{fn = list('sum' = list(na.rm = TRUE), 'sd')}.}
}
\value{
A tibble
}
\description{
This function provides a summary of a dataset, including both numeric and
non-numeric variables. For numeric variables, it calculates basic descriptive
statistics such as minimum, maximum, median, mean, and count of non-missing
values. Additionally, users can pass custom functions via the \code{fn} argument to
compute additional statistics for numeric variables. For non-numeric variables,
it provides frequency
counts and proportions for each unique value.
}
\examples{
mm_describe_df(data = data.frame(x = c(1:3, NA),
                                 y = c(3:4, NA, NA),
                                 z = c("A", "A", "B", "A")),
               y, x, z,
               fn = list('sum' = list(na.rm = TRUE), 'sd' = list(na.rm = TRUE))
              )

}
\seealso{
parse_list_fn
}
