% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_find_break.R
\name{mm_find_break}
\alias{mm_find_break}
\title{Detect time gaps in a datetime series}
\usage{
mm_find_break(data, datetime, format, threshold = 10, time_unit = "hours")
}
\arguments{
\item{data}{A data frame containing the datetime column.}

\item{datetime}{The datetime column.}

\item{format}{Optional. A character string specifying the datetime format, passed to \code{as.POSIXlt}.}

\item{threshold}{A numeric value indicating the minimum gap to be considered a break (default is 10).}

\item{time_unit}{The unit for the threshold. Supported values include "secs", "mins", "hours", "days", and "weeks".}
}
\value{
A tibble with columns \code{start}, \code{end}, and \code{duration}
showing the start and end of each break and its length.
}
\description{
Identifies breaks in a sequence of datetime observations based on a specified time threshold.
}
\examples{
pene <- read.csv(maimer:::table_files()[1])

set_cam <- pene \%>\%
  dplyr::filter(camera == "CAMERA 3") \%>\%
  dplyr::mutate(datetimes = as.POSIXlt(datetimes, format = "\%Y-\%m-\%d \%H:\%M:\%OS")) \%>\%
  dplyr::arrange(datetimes)

drtn <- mm_find_break(data = pene, datetime = "datetimes", threshold = 5, time_unit = "days")
drtn
}
