% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_fit_activity.R
\name{mm_fit_activity}
\alias{mm_fit_activity}
\title{Fit activity model to time-of-day data}
\usage{
mm_fit_activity(
  time_of_day,
  weights = NULL,
  n_boostrap = 1000,
  bandwidth = NULL,
  adjustment = 1,
  sample = c("none", "data", "model"),
  bounds = NULL,
  show = TRUE
)
}
\arguments{
\item{time_of_day}{A numeric vector of radian time-of-day data}

\item{weights}{A numeric vector of weights for each dat value.}

\item{n_boostrap}{Number of bootstrap iterations to perform. Ignored if sample=="none"}

\item{bandwidth}{Numeric value for kernel bandwidth. If NULL, calculated internally.}

\item{adjustment}{Numeric bandwidth adjustment multiplier.}

\item{sample}{Character string defining sampling method for bootstrapping errors (see details).}

\item{bounds}{A two-element vector defining radian bounds at which to truncate.}

\item{show}{Logical whether or not to show a progress bar while bootstrapping.}
}
\value{
A list
}
\description{
Fits kernel density to radian time-of-day data and estimates activity level from this distribution.
Optionally: 1. bootstraps the distribution, in which case SEs and confidence limits are also
stored for activity level and PDF; 2. weights the distribution; 3. truncates the distribution at given times.
}
\details{
When no \code{bounds} are given (default), a circular kernel distribution is fitted using \code{dvmkern}.
Otherwise, a normal kernel distribution is used, truncated at the values of \code{bounds}, using \code{density2}.

The bandwidth adjustment multiplier \code{adj} is provided to allow
exploration of the effect of adjusting the internally calculated bandwidth on
accuracy of activity level estimates.

The alternative bootstrapping methods defined by \code{sample} are:
\itemize{
 \item{\code{"none"}: no bootstrapping}
 \item{\code{"data"}: sample from the data}
 \item{\code{"model"}: sample from the fitted probability density distribution}
 }
It's generally better to sample from the data, but sampling from
the fitted distribution can sometimes provide more sensible confidence intervals when
the number of observations is very small.
}
