% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_independence.R
\name{mm_independence}
\alias{mm_independence}
\title{Evaluate independent detections}
\usage{
mm_independence(
  data = NULL,
  species_column,
  site_column,
  datetime,
  format,
  threshold = 30 * 60,
  only = FALSE
)
}
\arguments{
\item{data}{A \code{data.frame}, \code{tbl_df}, or \code{tbl} containing the event data. This should include
a column with datetime values. If \code{NULL}, the function will use the \code{datetime} argument
instead of the \code{data} argument.}

\item{species_column}{An optional column name specifying the species grouping.
If provided, independence will be assessed separately within each species group.}

\item{site_column}{An optional column name specifying the site/camera grouping.
If provided, independence will be assessed separately within each site group.}

\item{datetime}{A \code{character} string specifying the name of the column in \code{data} that contains
the datetime values. This argument is required if \code{data} is provided.}

\item{format}{A \code{character} string defining the format used to parse the datetime values in
the \code{datetime} column.}

\item{threshold}{A \code{numeric} value representing the time difference threshold (in seconds) to
determine whether events are independent. Events are considered independent if the time
difference between them is greater than or equal to this threshold. The default is 30 minutes
(1800 seconds).}

\item{only}{A \code{logical} value indicating whether to return only the rows of \code{data} that are
identified as independent events. If \code{TRUE}, only independent events are returned. If \code{FALSE},
the entire data frame is returned with an additional column indicating the independence status.
The default is \code{TRUE}.}
}
\value{
\itemize{
\item If \code{data} is provided and \code{only} is \code{TRUE}, a tibble of events identified as independent.
\item If \code{data} is provided and \code{only} is \code{FALSE}, a tibble of the original data with additional columns
indicating the \code{independent} status and \code{deltatime} differences (in second).
\item If \code{data} is not provided, a tibble of the \code{deltatime} values with \code{independent} status.
}
}
\description{
Filters camera trap data to ensure temporal independence between detections,
removing consecutive entry of the same species at the same location within a
specified time window.
}
\details{
Following Ridout & Linkie (2009), consecutive photos
of the same species at the same location within 30 minutes are considered non-independent and removed.

The approach mirrors the methodology applied by Linkie & Ridout (2011)
for Sumatran tiger-prey interactions study and Ahmad et al. (2024) to calculate
activity levels where such filtering is essential for:
\itemize{
\item Avoiding autocorrelation in activity pattern data
\item Ensuring each record represents an independent observation
\item Creating a random sample from the underlying activity distribution
}

The filtered data can then be used to estimate probability density functions of daily activity patterns,
assuming animals are equally detectable during their active periods.
}
\examples{

library(dplyr)

# Load example dataset
cam_data <- read.csv(system.file("penessoulou_season1.csv", package = "maimer"))

# Independence without considering species occurrence
indep1 <- cam_data \%>\%
  mm_independence(data = ., datetime = datetimes, format = "\%Y-\%m-\%d \%H:\%M:\%S",
                  only = TRUE)

sprintf("Independent observations: \%s", nrow(indep1))

# Independence considering species occurrence
indep2 <- cam_data \%>\%
  mm_independence(data = ., datetime = datetimes, format = "\%Y-\%m-\%d \%H:\%M:\%S",
                  only = TRUE, species_column = "species")

sprintf("Independent observations: \%s", nrow(indep2))

# Use a standalone vector of datetime values
dtime <- cam_data$datetimes
mm_independence(datetime = dtime, format = "\%Y-\%m-\%d \%H:\%M:\%S", only = TRUE)

}
\references{
Ridout, M.S., & Linkie, M. (2009). Estimating overlap of daily activity patterns
from camera trap data. Journal of Agricultural, Biological, and Environmental
Statistics, 14(3), 322-337. \url{https://doi.org/10.1198/jabes.2009.08038}

Linkie, M., & Ridout, M.S. (2011). Assessing tiger-prey interactions in Sumatran
rainforests. Journal of Zoology, 284(3), 224-229.
\url{https://doi.org/10.1111/j.1469-7998.2011.00801.x}

Ahmad, F., Mori, T., Rehan, M., Bosso, L., & Kabir, M. (2024). Applying a Random
Encounter Model to Estimate the Asiatic Black Bear (Ursus thibetanus) Density from
Camera Traps in the Hindu Raj Mountains, Pakistan. Biology, 13(5), 341.
\url{https://doi.org/10.3390/biology13050341}
}
