% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mm_lognorm_ci}
\alias{mm_lognorm_ci}
\title{Log-normal confidence interval}
\usage{
mm_lognorm_ci(estimate, se, percent = 95)
}
\arguments{
\item{estimate}{Numeric estimate value(s)}

\item{se}{Standard error(s) of the estimate}

\item{percent}{Percentage confidence level}
}
\value{
A dataframe with a row per estimate input, and columns \code{lcl}
and \code{ucl} (lower and upper confidence limits).
}
\description{
Calculates approximate log-normal confidence intervals given estimates
and their standard errors.
}
\examples{
  mm_lognorm_ci(10.13, 3.57)
}
