% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_overlap_estimates.R
\name{mm_overlap_estimates}
\alias{mm_overlap_estimates}
\title{Estimates of coefficient of overlapping}
\usage{
mm_overlap_estimates(
  A,
  B,
  kmax = 3,
  adjust = c(0.8, 1, 4),
  n_grid = 128,
  type = c("all", "Dhat1", "Dhat4", "Dhat5")
)
}
\arguments{
\item{A}{A numeric vector of time-of-day observations (in radians) for species A.}

\item{B}{A numeric vector of time-of-day observations (in radians) for species B.}

\item{kmax}{An integer indicating the maximum number of modes allowed in the activity pattern. Default is \code{3}.}

\item{adjust}{A numeric value to adjust the bandwidth of the kernel density estimation. Default is \code{1}.}

\item{n_grid}{An integer specifying the number of grid points for density estimation. Default is \code{128}.}

\item{type}{
the name of the estimator to use: \code{Dhat4} is recommended if both samples are larger then 50, otherwise use \code{Dhat1}. See Details. The default is "all" for compatibility with older versions.
}
}
\description{
Estimates of coefficient of overlapping
}
\examples{

set.seed(42)
species_A <- runif(100, 1.2, 2 * pi)
species_B <- runif(100, 0.23, 2 * pi)
mm_overlap_estimates(species_A, species_B)
mm_overlap_estimates(species_A, species_B, type = "Dhat4")

}
