% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_plot_density.R
\name{mm_plot_density}
\alias{mm_plot_density}
\title{Plot species' activity patterns}
\usage{
mm_plot_density(
  time,
  xscale = 24,
  xcenter = c("noon", "midnight"),
  n_grid = 128,
  kmax = 3,
  adjust = 1,
  rug = FALSE,
  linetype = 2,
  linecol = "gray10",
  linewidth = 1,
  rug_lentgh = 0.018,
  rug_color = "gray30",
  extend = "lightgrey",
  extend_alpha = 0.8,
  ...
)
}
\arguments{
\item{time}{A numeric vector of time-of-day observations (in radians, 0 to \eqn{2\pi}).}

\item{xscale}{A numeric value to scale the x-axis. Default is \code{24} for representing time in hours.}

\item{xcenter}{A string indicating the center of the x-axis. Options are \code{"noon"} (default) or \code{"midnight"}.}

\item{n_grid}{An integer specifying the number of grid points for density estimation. Default is \code{128}.}

\item{kmax}{An integer indicating the maximum number of modes allowed in the activity pattern. Default is \code{3}.}

\item{adjust}{A numeric value to adjust the bandwidth of the kernel density estimation. Default is \code{1}.}

\item{rug}{A logical value indicating whether to include a rug plot of the observations. Default is \code{FALSE}.}

\item{linetype}{A numeric specifying the line types. Default is \code{c(1, 2)}.}

\item{linecol}{A string specifying the colors of the density lines for species A and B. Default is \code{c("gray10", "gray0")}.}

\item{linewidth}{A numeric value specifying the line widths for species A and B density lines. Default is \code{c(1, 1)}.}

\item{rug_lentgh}{A numeric value specifying the length of the rug ticks. Default is \code{0.018} (in normalized plot coordinates).}

\item{rug_color}{A string specifying the color of the rug ticks. Default is \code{"gray30"}.}

\item{extend}{A string specifying the color of the extended area beyond the activity period. Default is \code{"lightgrey"}.}

\item{extend_alpha}{A numeric value (0 to 1) for the transparency of the extended area. Default is \code{0.8}.}

\item{...}{Additional arguments passed to the \code{geom_rug} function.}
}
\value{
A ggplot object representing the activity density curves of the species.
}
\description{
This function visualizes species' activity patterns based on time-of-day data.
It uses kernel density estimation to estimate activity density.
}
\examples{

 # Generate random data for two species
 set.seed(42)
 A <- runif(100, 0, 2 * pi)

 # Plot overlap with default settings
 mm_plot_density(A)
 # Customize plot with specific colors and line types
 mm_plot_density(A, linecol = "gray10", linewidth = 0.8,
                 xcenter = "midnight", rug = TRUE,
                 rug_color = 'red', extend_alpha = 0)



}
