% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_read.R
\name{mm_read}
\alias{mm_read}
\title{Read a Delimited File into a Tibble}
\usage{
mm_read(file_path, header = TRUE, sep, ...)
}
\arguments{
\item{file_path}{A string specifying the path to the file to be read.}

\item{header}{A logical value indicating whether the file contains a header row. Defaults to \code{TRUE}.}

\item{sep}{The field separator character. If not provided, the function automatically detects the separator.}

\item{...}{Additional arguments passed to the \code{read.table} function for fine-tuned control over file reading.}
}
\value{
A tibble containing the data from the specified file.
}
\description{
The \code{mm_read} function reads a delimited text file.
It automatically detects the delimiter if not specified and provides an easy-to-use
interface for importing data with additional customization options.
}
