% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_to_diel.R
\name{mm_to_diel}
\alias{mm_to_diel}
\title{Convert Time to Diel Fraction}
\usage{
mm_to_diel(times, sep)
}
\arguments{
\item{times}{A vector of time \code{as.POSIXct}, \code{strptime} or \code{strings}. Each string can represent a time in the format "HH:MM:SS" or variations thereof. The time may also include additional information, such as dates, which will be ignored in the conversion process.}

\item{sep}{A character string specifying the separator used in the \code{times} input. The separator could be ":", "-", or any character that separates hours, minutes, and seconds.}
}
\value{
A numeric vector of diel fractions, with each value ranging from 0 to 1, where:
\itemize{
\item \code{0} represents midnight (start of the day)
\item \code{0.5} represents noon (midday)
\item \code{1} represents the end of the day (just before the next midnight)
}
}
\description{
The \code{mm_to_diel} function converts a time string into a diel fraction,
which represents the time as a proportion of a 24-hour day. This is particularly
useful for analyzing timing and duration of activities throughout a day.
}
\details{
The function splits the \code{times} strings using the specified \code{sep} and any
whitespace, extracts the hours, minutes, and seconds, and then converts them
into a fraction of a 24-hour day (diel fraction). If more than three elements (e.g., "YYYY-MM-DD HH:MM:SS") are found, the function will only use the last three elements (assumed to be "HH:MM:SS").
}
\note{
The function assumes that the time strings are well-formed and will ignore empty or malformed components.
}
\examples{
# Example usage
times <- c("12:34:56", "23:45:00", "2024-08-18 06:30:15")
sep <- ":"
diel_times <- mm_to_diel(times, sep)
print(diel_times)

}
