% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_to_occupancy.R
\name{mm_to_occupancy}
\alias{mm_to_occupancy}
\title{Convert camera trap data to occupancy format}
\usage{
mm_to_occupancy(
  data,
  date_column,
  format = "\%Y-\%m-\%d",
  site_column,
  species_column,
  size_column,
  by_day = 7,
  presence_absence = TRUE
)
}
\arguments{
\item{data}{A data frame containing camera trap detection records.}

\item{date_column}{The name of the column containing detection dates.}

\item{format}{a character string. If not specified when converting from a
character representation, it will try c("\%Y-\%m-\%d", "\%Y/\%m/\%d") one by one,
and give an error if none works. Otherwise, the processing is
via \code{\link[=strptime]{strptime()}} whose help page describes available conversion specifications.}

\item{site_column}{The name of the column identifying sampling sites.}

\item{species_column}{The name of the column containing species names.
Can be NULL if species information is not needed.}

\item{size_column}{The name of the column representing detection counts.}

\item{by_day}{An integer specifying the number of days per time window (default: \code{7}).}

\item{presence_absence}{Logical. If \code{TRUE}, converts counts to presence-absence
data (1 = detected, 0 = not detected). Default is \code{TRUE}.}
}
\value{
A wide-format data frame where rows represent sites (and optionally species),
and columns represent detection windows. Values indicate either detection counts
or presence-absence (0/1).
}
\description{
This function transforms camera trap detection data into an occupancy format
suitable for analysis. It aggregates detections into user-defined time windows
and optionally converts counts into presence-absence (0/1) data.
}
\examples{

data <- data.frame(
  date = c("01-01-2023", "03-01-2023", "10-01-2023", "15-01-2023"),
  site = c("A", "A", "B", "B"),
  species = c("Tiger", "Tiger", "Deer", "Deer"),
  count = c(1, 2, 3, 1)
)

occupancy_data <- mm_to_occupancy(
  data,
  date_column = date,
  site_column = site,
  species_column = species,
  size_column = count,
  by_day = 7,
  presence_absence = TRUE
)

occupancy_data


}
\seealso{
\code{\link[=mm_to_community]{mm_to_community()}}
}
