% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_to_radian.R
\name{mm_to_radian}
\alias{mm_to_radian}
\title{Convert Time to Radians}
\usage{
mm_to_radian(data, times, format = "\%H:\%M:\%S", time_zone = "UTC")
}
\arguments{
\item{data}{A data frame containing a column with time values. Optional. If \code{NULL}, the \code{times} parameter is treated as a standalone vector.}

\item{times}{A column name in the \code{data} or a vector of time values to be converted. Time values should be in a format recognized by \code{as.POSIXct()}.}

\item{format}{A string specifying the format of the time values, using the standard POSIX formatting syntax. Default is \code{"\%H:\%M:\%S"}.}

\item{time_zone}{A string specifying the time zone for interpreting the time values. Default is \code{"UTC"}.}
}
\value{
If \code{data} is provided, the function returns the input data frame with an additional column named \code{time_radian}.
If \code{data} is not provided, the function returns a numeric vector of time values converted to radians.
}
\description{
This function converts time values into radians, which is often used in circular statistics and time-of-day analyses.
}
\details{
This function converts time values into radians based on a 24-hour clock:
\itemize{
\item A full day (24 hours) corresponds to \eqn{2\pi} radians.
\item The fractional time of the day is calculated as:
\deqn{\text{Fraction of the day} = \frac{\text{hours}}{24} + \frac{\text{minutes}}{1440} + \frac{\text{seconds}}{86400}}
}

For example, for a time of 23 hours, 6 minutes, and 12 seconds:
\deqn{\text{Fraction of the day} = \frac{23}{24} + \frac{6}{1440} + \frac{12}{86400}}

To convert this fraction into radians:
\deqn{\text{Radians} = \text{Fraction of the day} \times 2\pi}
}
\examples{
\dontrun{
  # Convert a standalone vector of time values
  times <- c("00:00:00", "06:00:00", "12:00:00", "18:00:00")
  mm_to_radian(times = times, format = "\%H:\%M:\%S")

  # Convert a column of time values in a data frame
  data <- data.frame(times = c("00:00:00", "06:00:00", "12:00:00", "18:00:00"))
  mm_to_radian(data = data, times = times, format = "\%H:\%M:\%S")
}

}
