% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_boot_CI.R
\name{mm_boot_CI}
\alias{mm_boot_CI}
\title{Bootstrap confidence intervals}
\usage{
mm_boot_CI(t0, bt, conf = 0.95)
}
\arguments{
\item{t0}{the statistic estimated from the original sample, usually the output from \code{\link[=mm_overlap_estimates]{mm_overlap_estimates()}}}

\item{bt}{a vector of bootstrap statistics, usually the output from \code{\link[=mm_boot_estimates]{mm_boot_estimates()}}}

\item{conf}{
a (single!) confidence interval to estimate.
}
}
\description{
Confidence interval calculation from bootstrap samples.
}
