% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_independence.R
\name{mm_independence}
\alias{mm_independence}
\title{Evaluate Event Independence}
\usage{
mm_independence(
  data = NULL,
  datetime,
  format,
  threshold = 30 * 60,
  only = FALSE
)
}
\arguments{
\item{data}{A \code{data.frame}, \code{tbl_df}, or \code{tbl} containing the event data. This should include
a column with datetime values. If \code{NULL}, the function will use the \code{deltatime} argument
instead of the \code{data} argument.}

\item{datetime}{A \code{character} string specifying the name of the column in \code{data} that contains
the datetime values. This argument is required if \code{data} is provided.}

\item{format}{A \code{character} string defining the format used to parse the datetime values in
the \code{datetime} column. This argument is required if \code{data} is provided and should be
in the format recognized by \code{\link[=strptime]{strptime()}}.}

\item{threshold}{A \code{numeric} value representing the time difference threshold (in seconds) to
determine whether events are independent. Events are considered independent if the time
difference between them is greater than or equal to this threshold. The default is 30 minutes
(1800 seconds).}

\item{only}{A \code{logical} value indicating whether to return only the rows of \code{data} that are
identified as independent events. If \code{TRUE}, only independent events are returned. If \code{FALSE},
the entire data frame is returned with an additional column indicating the independence status.
The default is \code{TRUE}.}

\item{deltatime}{A \code{numeric} vector of time differences between successive events. This argument
is used if \code{data} is not provided. If \code{data} is provided, this argument is ignored.}
}
\value{
\itemize{
\item If \code{data} is provided and \code{only} is \code{TRUE}, a tibble of events identified as independent.
\item If \code{data} is provided and \code{only} is \code{FALSE}, a tibble of the original data with additional columns
indicating the \code{independent} status and \code{deltatime} differences.
\item If \code{data} is not provided, a tibble of the \code{deltatime} values with \code{independent} status.
}
}
\description{
This function calculates the difference between times and evaluates whether
events are independent based on a given threshold. It is useful for checking
if certain events in a dataset are independent of each other based on time intervals.
}
\examples{
# Example with a data frame
df <- data.frame(datetime = as.POSIXct(c("2024-08-01 10:00:00", "2024-08-01 10:15:00",
                                         "2024-08-01 10:45:00", "2024-08-01 11:00:00")),
                 value = c(1, 2, 3, 4))
result <- mm_is_independent(data = df, datetime = "datetime", format = "\%Y-\%m-\%d \%H:\%M:\%S")

# Example with deltatime
deltatime <- c(1800, 1800, 3600)
result <- mm_is_independent(deltatime = deltatime)

}
