% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_solartime.R
\name{mm_solartime}
\alias{mm_solartime}
\title{Transform time to solar time anchored to sunrise and sunset}
\usage{
mm_solartime(
  data = NULL,
  date,
  longitude,
  latitude,
  crs = NULL,
  time_zone,
  ...,
  try_formats = c("\%Y-\%m-\%d \%H:\%M:\%OS", "\%Y/\%m/\%d \%H:\%M:\%OS",
    "\%Y:\%m:\%d \%H:\%M:\%OS", "\%Y-\%m-\%d \%H:\%M", "\%Y/\%m/\%d \%H:\%M",
    "\%Y:\%m:\%d \%H:\%M", "\%Y-\%m-\%d", "\%Y/\%m/\%d", "\%Y:\%m:\%d")
)
}
\arguments{
\item{data}{A data frame containing the date, longitude, and latitude columns.
If \code{NULL}, the function will use the \code{date}, \code{longitude}, and \code{latitude}
parameters directly. Default is \code{NULL}.}

\item{date}{A vector of date-time values or a column name in \code{data} representing
the date-time values to be converted to solar time. This can be a character
vector or a \code{POSIXlt} object.}

\item{longitude}{A numeric vector or a column name in \code{data} representing the
longitude of the location(s). Longitude should be in decimal degrees.}

\item{latitude}{A numeric vector or a column name in \code{data} representing the
latitude of the location(s). Latitude should be in decimal degrees.}

\item{crs}{A coordinate reference system (CRS) string or object specifying the
current CRS of the input coordinates. If provided, the function will transform
the coordinates to longitude and latitude (WGS84). This is useful when the input
coordinates are in a projected system (e.g., UTM). Default is \code{NULL}.}

\item{time_zone}{A numeric vector representing the time zone offset(s) from UTC
(in hours). If \code{data} is provided, this should match the number of unique
locations in the data.}

\item{...}{Additional arguments passed to \code{as.POSIXlt} for date parsing.}

\item{try_formats}{A vector of date-time formats to try when parsing the \code{date} parameter. Default includes common date-time formats.}
}
\value{
A tibble with the following columns:
\itemize{
\item \code{input}: The original date-time values.
\item \code{clock}: The local clock time.
\item \code{solar}: The calculated solar time.
}
If \code{data} is provided, the tibble will also include the longitude and latitude columns.
}
\description{
This function converts local time to solar time based on the sunrise and sunset
times for a given location. Solar time is a timekeeping system where the day
is defined by the position of the sun in the sky, with sunrise marking the start
of the day and sunset marking the end.
}
\details{
The function calculates solar time by first determining the sunrise and sunset times for the given location(s) and date(s). It then uses these times to anchor the solar time calculation. The solar time is computed by transforming the local clock time based on the position of the sun in the sky.

If \code{data} is provided, the function will process each unique location in the
data and return a tibble with the solar time for each date-time value. If \code{data}
is \code{NULL}, the function will process the \code{date}, \code{longitude}, and \code{latitude}
parameters directly.
}
\examples{
library(dplyr)

read.csv(system.file("penessoulou_season1.csv", package = "maimer")) \%>\%
 dplyr::filter(species == "Erythrocebus patas") \%>\%
 # Select independent events based on a given threshold
 maimer::mm_independence(species_column = species,
                          datetime = datetimes, threshold = 60*5, # 5 minutes
                          format = "\%Y-\%m-\%d \%H:\%M:\%S",
                          only = TRUE) \%>\%
 # Transform Time to Solar Time
 mm_solartime(data = ., date = datetime, longitude = longitude, latitude = latitude,
               crs = "EPSG:32631", time_zone = 1)

}
\references{
Rowcliffe, M. (2023). activity: Animal Activity Statistics.
R package version 1.3.4. https://CRAN.R-project.org/package=activity
}
